new analysis file to do a simple decision tree
For fun trying out what a decision tree will find looking at the data set. There is only a binary feature used for the dataset classification which may not be a great choice. I also changed the data input code to grab all the FPKM data for every file (so some case_id's are grabbed multiple times right now.) The decision tree analyzed the impact of every gene's FPKM to make a classification.